requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-sqlite==3.8.2
apispec==6.8.1
argcomplete==3.6.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
cattrs==24.1.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
email-validator==1.3.1
exceptiongroup==1.2.2
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozendict==2.4.6
frozenlist==1.5.0
google-re2==1.1.20240702
googleapis-common-protos==1.69.2
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.10.0
limits==4.7.3
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
more-itertools==10.6.0
msgspec==0.19.0
multidict==6.4.3
multitasking==0.0.11
numpy==1.24.4 
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-exporter-otlp-proto-http==1.32.0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
ordered-set==4.1.0
packaging==24.2
pandas==2.1.4
pathspec==0.12.1
peewee==3.17.9
pendulum==2.1.2
platformdirs==4.3.7
pluggy==1.5.0
prison==0.2.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.24.0
setproctitle==1.3.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
spark==0.2.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
text-unidecode==1.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
urllib3==2.4.0
Werkzeug==2.2.3
wrapt==1.17.2
WTForms==3.0.1
yarl==1.19.0
yfinance==0.2.31
zipp==3.21.0
